+++ src/types/trap.rs
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 //! Trap related types and utilities
 
+use core::fmt;
+
 /// Information about a trap
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub struct Info {
@@ -23,5 +25,15 @@ impl Info {
     /// This info refers to an exception
     pub fn is_exception(&self) -> bool {
         self.tval.is_some()
+    }
+}
+
+impl fmt::Display for Info {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let ecause = self.ecause;
+        match self.tval {
+            Some(tval) => write!(f, "exception (ecause: {ecause}, tval: {tval:0x})"),
+            None => write!(f, "interrupt (ecause: {ecause})"),
+        }
     }
 }